+++ codex/librarian/db/create_comics.py
@@ -4,7 +4,11 @@ from pathlib import Path
 from django.db.models import Q
 from django.db.models.functions import Now
 
-from codex.librarian.queue_mp import LIBRARIAN_QUEUE, BulkComicCoverCreateTask
+from codex.librarian.queue_mp import (
+    LIBRARIAN_QUEUE,
+    BulkComicCoverCreateTask,
+    CreateMissingCoversTask,
+)
 from codex.models import Comic, Credit, Folder, Imprint, Publisher, Series, Volume
 from codex.settings.logging import get_logger
 
@@ -91,6 +95,7 @@ def _update_comics(library, comic_paths, mds):
 
     LOG.verbose(f"Bulk updating {num_comics} comics.")
     count = Comic.objects.bulk_update(update_comics, BULK_UPDATE_COMIC_FIELDS)
+    LOG.verbose(f"Queueing cover regeneration for {len(comic_pks)} updated comics.")
     task = BulkComicCoverCreateTask(True, tuple(comic_pks))
     LIBRARIAN_QUEUE.put(task)
     return count
@@ -128,7 +133,8 @@ def _create_comics(library, comic_paths, mds):
     comic_pks = []
     for comic in created_comics:
         comic_pks.append(comic.pk)
-    task = BulkComicCoverCreateTask(True, tuple(comic_pks))
+    LOG.verbose(f"Queueing cover generation for {len(comic_pks)} created comics.")
+    task = BulkComicCoverCreateTask(False, tuple(comic_pks))
     LIBRARIAN_QUEUE.put(task)
     return num_comics
 
@@ -237,6 +243,8 @@ def bulk_import_comics(library, create_paths, update_paths, all_bulk_mds, all_m2
 
     update_count = _update_comics(library, update_paths, all_bulk_mds)
     create_count = _create_comics(library, create_paths, all_bulk_mds)
+    # Just to be sure.
+    LIBRARIAN_QUEUE.put(CreateMissingCoversTask())
 
     all_m2m_links = _link_comic_m2m_fields(all_m2m_mds)
     for field_name, m2m_links in all_m2m_links.items():