+++ codex/librarian/librariand.py
@@ -2,12 +2,16 @@
 from multiprocessing import Process
 from time import sleep
 
+from setproctitle import setproctitle
+
+from codex.darwin_mp import force_darwin_multiprocessing_fork
 from codex.librarian.covers.coverd import CoverCreator
 from codex.librarian.db.updaterd import Updater
 from codex.librarian.janitor.crond import Crond, janitor
 from codex.librarian.queue_mp import (
     LIBRARIAN_QUEUE,
     ComicCoverTask,
+    CreateMissingCoversTask,
     DelayedTasks,
     JanitorTask,
     NotifierTask,
@@ -27,6 +31,7 @@ from codex.librarian.watchdog.observers import (
 from codex.notifier import Notifier
 from codex.settings.logging import get_logger
 from codex.threads import QueuedThread
+from codex.version import PACKAGE_NAME
 
 
 LOG = get_logger(__name__)
@@ -54,7 +59,7 @@ class LibrarianDaemon(Process):
     proc = None
 
     def __init__(self):
-        """Create threads and thread pool."""
+        """Init process."""
         super().__init__(name=self.NAME, daemon=False)
 
     def _process_task(self, task):
@@ -88,6 +93,7 @@ class LibrarianDaemon(Process):
 
     def _create_threads(self):
         """Create all the threads."""
+        force_darwin_multiprocessing_fork()
         self.delayed_tasks = DelayedTasksThread()
         self.cover_creator = CoverCreator()
         self.search_indexer = SearchIndexer()
@@ -133,6 +139,7 @@ class LibrarianDaemon(Process):
         threads.
         """
         try:
+            setproctitle(f"{PACKAGE_NAME}-{self.NAME}")
             LOG.verbose("Started Librarian process.")
             self._create_threads()
             self._start_threads()
@@ -159,6 +166,7 @@ class LibrarianDaemon(Process):
         cls.proc = LibrarianDaemon()
         cls.proc.start()
         LIBRARIAN_QUEUE.put(WatchdogSyncTask())
+        LIBRARIAN_QUEUE.put(CreateMissingCoversTask())
 
     @classmethod
     def shutdown(cls):